fixes docs compilation error
